outsources mini map rendering into uiGenerator
