+++ game.js
@@ -8,6 +8,7 @@ process.on('uncaughtException', function (exception) {
 });
 var LocalStorage = require('node-localstorage').LocalStorage;
 var HttpWrapper = require("./httpWrapper");
+const uiGenerator = require("./uiGenerator");
 const pngUtil = require("./pngUtil");
 const PNG = require('pngjs').PNG;
 localStorage = new LocalStorage('./localStorage');
@@ -291,22 +292,7 @@ Game.prototype.init = function () {
         }
     });
     socket.on("start", function () {
-        function clamp(x, low, high) {
-            return Math.min(Math.max(x, low), high);
-        }
 
-        function bSearch(search, arr) {
-            let low = 0, high = arr.length - 1;
-            while (low + 1 !== high) {
-                let mid = Math.floor((low + high) / 2);
-                if (arr[mid][0] >= search) {
-                    high = mid;
-                } else {
-                    low = mid;
-                }
-            }
-            return high;
-        }
 
         var hitLog = [];
         socket.on("hit", function (data) {
@@ -315,99 +301,8 @@ Game.prototype.init = function () {
 
         setTimeout(function () {
             setInterval(function () {
-                //Minimap creation
-                var png = new PNG({
-                    width: 188 * 2,
-                    height: 100 * 2,
-                    filterType: -1
-                });
-                for (let j = 0; j < Math.floor(png.data.length / 4); j++) {
-                    let idx = j << 2;
-                    png.data[idx] = 5;
-                    png.data[idx + 1] = 0;
-                    png.data[idx + 2] = 0;
-                    png.data[idx + 3] = 255;
-                }
-                var screenSize = {width: 800, height: 600};
-                var pos = {
-                    x: character.real_x - Math.floor(screenSize.width / 2),
-                    y: character.real_y - Math.floor(screenSize.height / 2)
-                };
-                var map = character.in;
-
-                if (G.maps[map]) {
-                    let xLines = G.maps[map].data.x_lines;
-                    let yLines = G.maps[map].data.y_lines;
-
-                    for (let i = bSearch(pos.x, xLines); i < xLines.length && xLines[i][0] < pos.x + screenSize.width; i++) {
-                        let line = xLines[i];
-                        let x = ((line[0] - pos.x) / screenSize.width) * png.width;
-                        let y1 = ((line[1] - pos.y) / screenSize.height) * png.height;
-                        let y2 = ((line[2] - pos.y) / screenSize.height) * png.height;
-                        pngUtil.draw_line(x, y1, x, y2, png, [255, 255, 255, 255]);
-                    }
-
-                    for (let i = bSearch(pos.y, yLines); i < yLines.length && yLines[i][0] < pos.y + screenSize.height; i++) {
-                        let line = yLines[i];
-                        let y = ((line[0] - pos.y) / screenSize.height) * png.height;
-                        let x1 = ((line[1] - pos.x) / screenSize.width) * png.width;
-                        let x2 = ((line[2] - pos.x) / screenSize.width) * png.width;
-                        pngUtil.draw_line(x1, y, x2, y, png, [255, 255, 255, 255]);
-                    }
-                }
-                // draw hit lines
-                for (let hit of hitLog) {
-                    let source = entities[hit.hid];
-                    let target = entities[hit.id];
-                    if (hit.hid === character.id)
-                        source = character;
-                    if (hit.id === character.id)
-                        target = character;
-
-                    if (source && target) {
-                        let color = [255, 255, 255, 255];
-                        if (hit.anim === "heal") {
-                            color = [0, 250, 0, 255];
-                        } else
-                            color = [250, 0, 0, 255];
-                        pngUtil.draw_line(
-                            ((source.real_x - pos.x) / screenSize.width) * png.width,
-                            ((source.real_y - pos.y) / screenSize.height) * png.height,
-                            ((target.real_x - pos.x) / screenSize.width) * png.width,
-                            ((target.real_y - pos.y) / screenSize.height) * png.height,
-                            png, [200, 0, 0, 255]);
-                    }
-                }
-                hitLog = [];
-                //draw entities
-                if (entities) {
-                    for (let id in entities) {
-                        let entity = entities[id];
-                        let color = [255, 255, 255, 255];
-                        if (entity.type === "monster") {
-                            if (!entity.dead) {
-                                color = [200, 0, 0, 255];
-                            } else {
-                                color = [100, 100, 100, 255];
-                            }
-                        }
-                        pngUtil.draw_dot(
-                            ((entity.real_x - pos.x) / screenSize.width) * png.width,
-                            ((entity.real_y - pos.y) / screenSize.height) * png.height,
-                            2,
-                            png, color);
-                    }
-                }
-                //draw character
-                pngUtil.draw_dot(
-                    ((character.real_x - pos.x) / screenSize.width) * png.width,
-                    ((character.real_y - pos.y) / screenSize.height) * png.height,
-                    2,
-                    png, [0, 200, 0, 255]);
-
-
-                let buffer = PNG.sync.write(png, options);
-
+                let buffer = uiGenerator.generateMiniMap(hitLog,entities);
+                hitLog= [];
                 process.send({
                     type: "bwiPush",
                     name: "minimap",
@@ -473,26 +368,6 @@ Game.prototype.init = function () {
                     }
                 }
 
-                var targetIds = [];
-                if(character.target)
-                    targetIds.push(character.target);
-                for (let id in entities) {
-                    let entity = entities[id];
-                    if (entity.target === character.id)
-                        if(!targetIds.includes(entity.id))
-                            targetIds.push(entity.id);
-                }
-
-                var targets = [];
-                for(let id of targetIds){
-                    if(entities[id])
-                        targets.push({
-                            name: entities[id].name,
-                            health: Math.floor(entities[id].hp * 10000 / entities[id].max_hp) / 100,
-                            level: entities[id].level,
-                        })
-
-                }
 
                 process.send({
                     type: "bwiUpdate",
@@ -503,7 +378,6 @@ Game.prototype.init = function () {
                         xp: Math.floor(character.xp * 10000 / character.max_xp) / 100,
                         health: Math.floor(character.hp * 10000 / character.max_hp) / 100,
                         mana: Math.floor(character.mp * 10000 / character.max_mp) / 100,
-                        targets: targets,
                         status: character.rip ? "Dead" : "Alive",
                         gold: toPrettyNum(character.gold),
                         dps: Math.floor(dps),
