+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "adventurelandbot",
-  "version": "1.5.1",
+  "version": "1.5.2",
   "description": "",
   "main": "main.js",
   "scripts": {
